tests/wpt/tests/html/semantics/embedded-content/media-elements/track/track-element/track-webvtt-empty-cue.html

<!DOCTYPE html>
<title>Empty cues should not be discarded</title>
<script src="track-helpers.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
check_cues_from_track("resources/empty-cue.vtt", function(track) {
    assert_equals(track.cues.length, 3);
});
</script>
